Login Page added
This is what I have for the login page as of tonight. Also added folders
css and img for holding css and images I used.

A lot of this will be cleaned up and made to properly work with
everything else, this is just where it stands as of tonight.