Updates. Please read description.
login.hjs:
Updated to show the post formatting and to accept the post.css
stylesheet
Updated to work with old form handler. (I plan to update this handler to
john's version as I like it better.

login.css:
Small style changes to fit with slight design changes. Still subject to
changing but that is easy to do.
Removed every reference to posts as that is now handled by the
specialized post.css file. Posts are a common element and should have
consistency.

login.js:
Updated datepicker in registration form to disallow future dates to be
chosen
Includes old form handler which I plan on converting to John's form
handler later

postImg.png:
Placeholder used to show image used in posts