Loading post functionality script
\public\temps:
Added a template folder to public. This was the only way I could reach
the template I needed using jQuery. If this functionality is moved to
server side with a simple call from the browser this will not be needed
any more.

login.hjs:
Now includes a simple jQuery script to load posts iteratively. Does not
render through hogan and express however, script will need to be updated
or this will have to be done server side. To load many posts iteratively
it is possible to use a nested JSON. See:
http://stackoverflow.com/questions/29998595/for-loop-in-hogan-js-template

postTemp.hjs:
The template html for posts has been moved outside of login into it's
own file so it may be easily accessed from every page. May have to
change how the comments sections currently work.